requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==2.0.4
djangorestframework==3.8.1
djangorestframework-jwt==1.11.0
kombu==4.1.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.3
redis==2.10.6
selenium==3.11.0
vine==1.1.4
